Water Logging Database
After a user signs in with a google account a document is created where the information of the user daily water intake is going to be saved. after the user opens the water logging screen a new array is created if there isn't one for the day. pressing the button adds 250 ml to the database and updates the information in the app from the database