Water Intake Logging
Now it saves the water intake data in the path "users/currentUsers.UID" in a array "Water Intake". The array holds a map for each day containing the date and amount drunk. Water Intake Screen shows todays drunk amount, yesterdays amount in ml, this weeks and this months amount in liters