Database changes
Changed it so that the water intake data would be saved in "journal/$UID/hydration/waterIntake".
Created a function that creates the $UID document with the hydration acticity collections with waterIntake doc in hydration collection and steps doc in activity collection. Fixed an issue where after changing acounts it would display and add water to the previous account